pi-requirements.txt
python-socketio[client]==5.10.0
